+++ pkg/installation/info.go
@@ -20,10 +20,10 @@ type installationInfo struct {
 func getInstallationInfo(httpClient *http.Client, apiUrl string) (installationInfo, error) {
 	res, err := httpClient.Get(apiUrl) // #nosec G107
 	if err != nil {
-		return installationInfo{}, microerror.Maskf(cannotGetInstallationInfo, "make sure you're connected to the internet and that the Giant Swarm API is up and running\n%v", err.Error())
+		return installationInfo{}, microerror.Maskf(cannotGetInstallationInfoError, "make sure you're connected to the internet and that the Giant Swarm API is up and running\n%s", err.Error())
 	}
 	if res.StatusCode != http.StatusOK {
-		return installationInfo{}, microerror.Maskf(cannotGetInstallationInfo, "make sure you're behind the correct VPN")
+		return installationInfo{}, microerror.Maskf(cannotGetInstallationInfoError, "make sure you're behind the correct VPN")
 	}
 
 	defer res.Body.Close()
@@ -32,7 +32,7 @@ func getInstallationInfo(httpClient *http.Client, apiUrl string) (installationIn
 	{
 		err = json.NewDecoder(res.Body).Decode(&result)
 		if err != nil {
-			return installationInfo{}, microerror.Maskf(cannotGetInstallationInfo, "API response has invalid format")
+			return installationInfo{}, microerror.Maskf(cannotGetInstallationInfoError, "API response has invalid format")
 		}
 
 		result.Installation.K8sCaCert, err = parseCertificate(result.Installation.K8sCaCert)