+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ## [Unreleased]
 
 - Removed the `--credential` flag, now it is managed by admission controller.
+- Add support for fetching installation information using non-standard Giant Swarm API URLs.
 
 ## [0.14.0] - 2020-11-24
 
@@ -18,7 +19,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Removed
 
-- Remove client-side validation of the GS `release` when creating a `Cluster`'s template. 
+- Remove client-side validation of the GS `release` when creating a `Cluster`'s template.
 
 
 ## [0.13.0] - 2020-11-20