+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [0.15.0] - 2020-12-07
+
 - Removed the `--credential` flag, now it is managed by admission controller.
 - Add support for fetching installation information using non-standard Giant Swarm API URLs.
 
@@ -209,7 +211,8 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v0.14.0...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v0.15.0...HEAD
+[0.15.0]: https://github.com/giantswarm/kubectl-gs/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/giantswarm/kubectl-gs/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/giantswarm/kubectl-gs/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/giantswarm/kubectl-gs/compare/v0.11.0...v0.12.0